Update wrapper for Notepad==
